lp_health_tracker/requirements.txt
# LP Health Tracker - Dependencies

# Web3 and blockchain interaction
web3==6.11.3
eth-account==0.9.0

# HTTP requests and API interactions
requests==2.31.0
aiohttp==3.9.1

# Environment variables
python-dotenv==1.0.0

# Task scheduling
APScheduler==3.10.4

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# YAML configuration
PyYAML>=6.0.0

# Data handling (updated for Python 3.13 compatibility)
pandas>=2.2.0
numpy>=1.26.0

# Database (SQLite is built-in Python module)
# sqlite3  # Built-in Python module - no installation needed

# Logging and utilities
colorlog==6.8.0

# Math and calculations (for IL formulas) - optional for MVP
# scipy>=1.13.0  # Commented out for now
# matplotlib>=3.8.0  # Commented out for now

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# JSON handling (built-in)
# datetime (built-in)
# logging (built-in)
